Move TxPublish to separate file
